requirements.txt
torch
transformers
accelerate
tqdm
numpy
pandas
pillow
matplotlib
scikit-learn
opencv-python
einops
timm
pycocotools
